pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "football_analytics"
version = "0.1.0"
description = "Football analytics library"
authors = [
  { name = "AI Ballers" },
]

readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",

]
requires-python = ">=3.12"

dependencies = [
    "black",
    "ruff",
    "mypy",
    "opencv-python",
    "httpx",
    "fastapi",
    "flake8",
    "isort",
    "loguru",
    "mkdocs",
    "pip",
    "types-requests",
    "python-dotenv",
    "typer",
    "ultralytics",
    "uvicorn",
    "supervision",
    "requests",
    "pytest",
    "pytest-cov",
    "python-multipart",
    "kagglehub",
    "tensorboard",
    "grpcio>=1.69.0",
    "grpcio.tools>=1.69.0",
    "grpc-stubs",
    "multimethod",
    "optuna",
    "grpcio>=1.69.0",
    "grpcio-tools>=1.69.0",
    "transformers[torch]",
    "sentencepiece",
    "more-itertools",
    "joblib==1.4.2",
    "llvmlite==0.43.0",
    "numba==0.60.0",
    "numpy==2.0.2",
    "pynndescent==0.5.13",
    "scikit-learn==1.6.1",
    "scipy==1.15.1",
    "threadpoolctl==3.5.0",
    "tqdm==4.67.1",
    "umap-learn==0.5.7"
  ]


[tool.ruff]
line-length = 99
exclude = ["*/grpc_files/", "notebooks/*"]

[tool.ruff.lint]
extend-select = ["E501"]
select = ["I", "F"]

[tool.pytest.ini_options]
pythonpath = ["football_analytics"]
addopts = "--cov=football_analytics --cov-report=term-missing"

[tool.mypy]
check_untyped_defs = true
ignore_missing_imports = true
exclude = ["grpc_files"]
